pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Neel Gokhale <neel@cohere.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cohere = "^5.9.1"
langchain-cohere = "^0.2.4"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"
shap = "^0.46.0"
scikit-learn = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
